reader-function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore